:recycle: Move some filter tests into the correct test module
These were smeared out over two test modules.
