+++ src/openzaak/components/zaken/tests/test_filters.py
@@ -4,7 +4,11 @@ from django.test import tag
 
 from rest_framework import status
 from rest_framework.test import APITestCase
-from vng_api_common.constants import RolOmschrijving, RolTypes
+from vng_api_common.constants import (
+    RolOmschrijving,
+    RolTypes,
+    VertrouwelijkheidsAanduiding,
+)
 from vng_api_common.tests import get_validation_errors, reverse
 
 from openzaak.utils.tests import JWTAuthMixin
@@ -206,8 +210,118 @@ class ZaakFilterTests(JWTAuthMixin, APITestCase):
         response = self.client.get(
             reverse(Zaak),
             {"rol__omschrijvingGeneriek": "behandelaar"},
-            **ZAAK_READ_KWARGS
+            **ZAAK_READ_KWARGS,
         )
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.json()["count"], 1)
+
+    def test_filter_startdatum(self):
+        ZaakFactory.create(startdatum="2019-01-01")
+        ZaakFactory.create(startdatum="2019-03-01")
+        url = reverse("zaak-list")
+
+        response_gt = self.client.get(
+            url, {"startdatum__gt": "2019-02-01"}, **ZAAK_READ_KWARGS
+        )
+        response_lt = self.client.get(
+            url, {"startdatum__lt": "2019-02-01"}, **ZAAK_READ_KWARGS
+        )
+        response_gte = self.client.get(
+            url, {"startdatum__gte": "2019-03-01"}, **ZAAK_READ_KWARGS
+        )
+        response_lte = self.client.get(
+            url, {"startdatum__lte": "2019-01-01"}, **ZAAK_READ_KWARGS
+        )
+
+        for response in [response_gt, response_lt, response_gte, response_lte]:
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            self.assertEqual(response.data["count"], 1)
+
+        self.assertEqual(response_gt.data["results"][0]["startdatum"], "2019-03-01")
+        self.assertEqual(response_lt.data["results"][0]["startdatum"], "2019-01-01")
+        self.assertEqual(response_gte.data["results"][0]["startdatum"], "2019-03-01")
+        self.assertEqual(response_lte.data["results"][0]["startdatum"], "2019-01-01")
+
+    def test_sort_startdatum(self):
+        ZaakFactory.create(startdatum="2019-01-01")
+        ZaakFactory.create(startdatum="2019-03-01")
+        ZaakFactory.create(startdatum="2019-02-01")
+        url = reverse("zaak-list")
+
+        response = self.client.get(url, {"ordering": "-startdatum"}, **ZAAK_READ_KWARGS)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.data["results"]
+
+        self.assertEqual(data[0]["startdatum"], "2019-03-01")
+        self.assertEqual(data[1]["startdatum"], "2019-02-01")
+        self.assertEqual(data[2]["startdatum"], "2019-01-01")
+
+    def test_filter_max_vertrouwelijkheidaanduiding(self):
+        zaak1 = ZaakFactory.create(
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.openbaar
+        )
+        zaak2 = ZaakFactory.create(
+            vertrouwelijkheidaanduiding=VertrouwelijkheidsAanduiding.geheim
+        )
+
+        url = reverse(Zaak)
+
+        response = self.client.get(
+            url,
+            {
+                "maximaleVertrouwelijkheidaanduiding": VertrouwelijkheidsAanduiding.zaakvertrouwelijk
+            },
+            **ZAAK_READ_KWARGS,
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["count"], 1)
+        self.assertEqual(
+            response.data["results"][0]["url"], f"http://testserver{reverse(zaak1)}",
+        )
+        self.assertNotEqual(
+            response.data["results"][0]["url"], f"http://testserver{reverse(zaak2)}",
+        )
+
+    def test_filter_rol__betrokkeneIdentificatie__natuurlijkPersoon__inpBsn_max_length(
+        self,
+    ):
+        ZaakFactory.create(startdatum="2019-01-01")
+        ZaakFactory.create(startdatum="2019-03-01")
+        url = reverse("zaak-list")
+
+        response = self.client.get(
+            url,
+            {"rol__betrokkeneIdentificatie__natuurlijkPersoon__inpBsn": "0" * 10},
+            **ZAAK_READ_KWARGS,
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        error = get_validation_errors(
+            response, "rol__betrokkeneIdentificatie__natuurlijkPersoon__inpBsn"
+        )
+        self.assertEqual(error["code"], "max_length")
+
+    def test_filter_rol__betrokkeneIdentificatie__medewerker__identificatie_max_length(
+        self,
+    ):
+        ZaakFactory.create(startdatum="2019-01-01")
+        ZaakFactory.create(startdatum="2019-03-01")
+        url = reverse("zaak-list")
+
+        response = self.client.get(
+            url,
+            {"rol__betrokkeneIdentificatie__medewerker__identificatie": "0" * 25},
+            **ZAAK_READ_KWARGS,
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        error = get_validation_errors(
+            response, "rol__betrokkeneIdentificatie__medewerker__identificatie"
+        )
+        self.assertEqual(error["code"], "max_length")
